--- subgraphs/seed-of-life/src/helpers/RealmHelpers.ts
@@ -1,14 +1,0 @@
-import { RealmStat } from "../../generated/schema";
-
-export class RealmStatHelpers {
-  public static getOrCreateRealmStat(name: string): RealmStat {
-    let realmStat = RealmStat.load(name);
-    if (!realmStat) {
-      realmStat = new RealmStat(name);
-      realmStat.realm = name;
-      realmStat.save();
-    }
-
-    return realmStat;
-  }
-}
