+++ subgraphs/seed-of-life/src/helpers/user-approval.ts
@@ -1,0 +1,5 @@
+import { Approval, User } from "../../generated/schema";
+
+export function getUserApprovalId(user: User, approval: Approval): string {
+  return `${user.id}-${approval.id}`;
+}
