+++ subgraphs/seed-of-life/src/helpers/user.ts
@@ -1,0 +1,11 @@
+import { User } from "../../generated/schema";
+
+export function getOrCreateUser(id: string): User {
+  let user = User.load(id);
+  if (!user) {
+    user = new User(id);
+    user.save();
+  }
+
+  return user;
+}
