+++ subgraphs/seed-of-life/src/mappings/balancer-crystal.ts
@@ -6,10 +6,10 @@ import {
   TransferSingle,
 } from "../../generated/BalancerCrystal/ERC1155";
 import { UserApproval } from "../../generated/schema";
-import { ApprovalHelpers } from "../helpers/ApprovalHelpers";
-import { TransferHelpers } from "../helpers/TransferHelpers";
-import { UserApprovalHelpers } from "../helpers/UserApprovalHelpers";
-import { UserHelpers } from "../helpers/UserHelpers";
+import { getOrCreateApproval } from "../helpers/approval";
+import { handleTransfer } from "../helpers/transfer";
+import { getOrCreateUser } from "../helpers/user";
+import { getUserApprovalId } from "../helpers/user-approval";
 
 export function handleTransferBatch(event: TransferBatch): void {
   let params = event.params;
@@ -17,7 +17,7 @@ export function handleTransferBatch(event: TransferBatch): void {
   for (let index = 0; index < params.ids.length; index++) {
     let id = params.ids[index];
 
-    TransferHelpers.handleTransfer(
+    handleTransfer(
       event.address,
       params.from,
       params.to,
@@ -29,7 +29,7 @@ export function handleTransferBatch(event: TransferBatch): void {
 }
 
 export function handleTransferSingle(event: TransferSingle): void {
-  TransferHelpers.handleTransfer(
+  handleTransfer(
     event.address,
     event.params.from,
     event.params.to,
@@ -40,14 +40,14 @@ export function handleTransferSingle(event: TransferSingle): void {
 }
 
 export function handleApprovalForAll(event: ApprovalForAll): void {
-  let user = UserHelpers.getOrCreateUser(event.params.account.toHexString());
+  let user = getOrCreateUser(event.params.account.toHexString());
 
   let contract = event.address;
   let operator = event.params.operator;
 
-  let approval = ApprovalHelpers.getOrCreateApproval(contract, operator);
+  let approval = getOrCreateApproval(contract, operator);
 
-  let userApprovalId = UserApprovalHelpers.getUserApprovalId(user, approval);
+  let userApprovalId = getUserApprovalId(user, approval);
 
   if (event.params.approved) {
     let userApproval = new UserApproval(userApprovalId);
