+++ subgraphs/seed-of-life/src/mappings/magic.ts
@@ -1,18 +1,18 @@
 import { Approval as ApprovalEvent } from "../../generated/Magic/ERC20";
-import { Approval, User, UserApproval } from "../../generated/schema";
-import { ApprovalHelpers } from "../helpers/ApprovalHelpers";
-import { UserApprovalHelpers } from "../helpers/UserApprovalHelpers";
-import { UserHelpers } from "../helpers/UserHelpers";
+import { UserApproval } from "../../generated/schema";
+import { getOrCreateApproval } from "../helpers/approval";
+import { getOrCreateUser } from "../helpers/user";
+import { getUserApprovalId } from "../helpers/user-approval";
 
 export function handleApproval(event: ApprovalEvent): void {
-  let user = UserHelpers.getOrCreateUser(event.params.owner.toHexString());
+  let user = getOrCreateUser(event.params.owner.toHexString());
 
   let contract = event.address;
   let spender = event.params.spender;
 
-  let approval = ApprovalHelpers.getOrCreateApproval(contract, spender);
+  let approval = getOrCreateApproval(contract, spender);
 
-  let userApprovalId = UserApprovalHelpers.getUserApprovalId(user, approval);
+  let userApprovalId = getUserApprovalId(user, approval);
 
   if (UserApproval.load(userApprovalId)) {
     return;
